refactor: consolidate state module
